Move line parsing to its own module
